Formatting fixes for PrivacyCursor
Removed tabs, fixed comments as well as some nasty if/then statements not lining up and really
crushing readability. To the extent that I hope I did not makes things worse and that it's
hard to tell. As before, I kept this separate from other merge commits so wbedard can revert
and start from scratch as needed.