Standardize formatting for PhoneStateListener.java
Removed wayward debugging addition as well. There are two 'BEGIN PRIVACY' sections here
separated only by the name of the function. Part of me thinks that it would be easier to
make that one hunk while the other thinks it is more likely to misapply as one hunk.

Something to consider.