SystemProperties: Default-Deny (#4)
Changed the implementation of this method to the OPD version found in Camera and elsewhere.
Since the method from CollegeDev was c/p and I assume it was rewritten for a reason, I don't
see why we don't implement it everywhere the original method was pasted.

The default-deny is using CollegeDev's implementation, which will be made very clear in the
squashed commit that will go into the -devel branch when ready.