Changing ERROR back to return code 4 in prep for PDroid2.0 merge
Used in handling default-deny object and changed for PDroid22.0/PdroidManager interchangablilty in mind.
The change to 1 was for backwards compatability, which is now at the very bottom of my personal list.
If it is up to me to backwards maintain, older Android versions will probably have to make do with the
currently available PDroid options, which are manifold.